improve a bit config

+++ update_symbolic_links.sh
@@ -15,7 +15,6 @@ ln -sf $basedir/terminal/kitty.conf $HOME/.config/kitty/
 mkdir -p $HOME/.config/alacritty
 ln -sf $basedir/terminal/alacritty.yml $HOME/.config/alacritty
 
-
 # The config dir
 mkdir -p $HOME/.config/clash
 ln -sf $basedir/config/clash.yaml $HOME/.config/clash
@@ -24,7 +23,13 @@ mkdir -p $HOME/.config/fcitx5
 ln -sf $basedir/config/fcitx5.conf $HOME/.config/fcitx5/config
 ln -sf $basedir/config/fcitx5.profile $HOME/.config/fcitx5/profile
 
-sudo cp $basedir/config/ly /etc/pam.d
+if [[ ! -f /etc/pam.d/ly ]]; then
+	echo "Install ly pam config"
+	sudo cp $basedir/config/ly /etc/pam.d
+elif ! cmp $basedir/config/ly /etc/pam.d/ly; then
+	echo "Restore ly pam config"
+	sudo cp $basedir/config/ly /etc/pam.d
+fi
 
 if [[ -d $HOME/Documents/Project/qutebrowser ]]; then
 	mkdir -p $HOME/.local/bin
@@ -33,11 +38,16 @@ fi
 mkdir -p $HOME/.config/qutebrowser
 ln -sf $basedir/config/qutebrowser-config.py $HOME/.config/qutebrowser/config.py
 
-sudo cp $basedir/config/wifi_login /etc/NetworkManager/dispatcher.d
+if [[ ! -f /etc/NetworkManager/dispatcher.d/wifi_login ]]; then
+	echo "Install wifi_login hook"
+	sudo cp $basedir/config/wifi_login /etc/NetworkManager/dispatcher.d
+elif ! cmp $basedir/config/wifi_login /etc/NetworkManager/dispatcher.d/wifi_login; then
+	echo "Restore wifi_login hook"
+	sudo cp $basedir/config/wifi_login /etc/NetworkManager/dispatcher.d
+fi
 
 mkdir -p $HOME/.config/zathura
 ln -sf $basedir/config/zathurac $HOME/.config/zathura
-
 
 # Config sway
 mkdir -p $HOME/.config/sway
